move processors init out of main func
