+++ metrics/processors/factory.go
@@ -1,0 +1,40 @@
+// Copyright 2015 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package processors
+
+import (
+	"fmt"
+
+	"k8s.io/heapster/common/flags"
+	"k8s.io/heapster/metrics/core"
+	"k8s.io/heapster/metrics/processors/kubeprocessors"
+)
+
+type ProcessorFactory struct {
+}
+
+func (this *ProcessorFactory) Build(uri flags.Uri) ([]core.DataProcessor, error) {
+	switch uri.Key {
+	case "kubernetes":
+		processors, err := kubeprocessors.GetKubeProcessors(&uri.Val)
+		return processors, err
+	default:
+		return nil, fmt.Errorf("Processor not recognized: %s", uri.Key)
+	}
+}
+
+func NewProcessorFactory() *ProcessorFactory {
+	return &ProcessorFactory{}
+}
