+++ metrics/processors/kubeprocessors/cluster_aggregator_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package processors
+package kubeprocessors
 
 import (
 	"testing"
@@ -23,14 +23,13 @@ import (
 	"k8s.io/heapster/metrics/core"
 )
 
-func TestPodAggregator(t *testing.T) {
+func TestClusterAggregate(t *testing.T) {
 	batch := core.DataBatch{
 		Timestamp: time.Now(),
 		MetricSets: map[string]*core.MetricSet{
-			core.PodContainerKey("ns1", "pod1", "c1"): {
+			core.PodKey("ns1", "pod1"): {
 				Labels: map[string]string{
-					core.LabelMetricSetType.Key: core.MetricSetTypePodContainer,
-					core.LabelPodName.Key:       "pod1",
+					core.LabelMetricSetType.Key: core.MetricSetTypeNamespace,
 					core.LabelNamespaceName.Key: "ns1",
 				},
 				MetricValues: map[string]core.MetricValue{
@@ -47,10 +46,9 @@ func TestPodAggregator(t *testing.T) {
 				},
 			},
 
-			core.PodContainerKey("ns1", "pod1", "c2"): {
+			core.PodKey("ns1", "pod2"): {
 				Labels: map[string]string{
-					core.LabelMetricSetType.Key: core.MetricSetTypePodContainer,
-					core.LabelPodName.Key:       "pod1",
+					core.LabelMetricSetType.Key: core.MetricSetTypeNamespace,
 					core.LabelNamespaceName.Key: "ns1",
 				},
 				MetricValues: map[string]core.MetricValue{
@@ -68,29 +66,19 @@ func TestPodAggregator(t *testing.T) {
 			},
 		},
 	}
-	processor := PodAggregator{}
+	processor := ClusterAggregator{
+		MetricsToAggregate: []string{"m1", "m3"},
+	}
 	result, err := processor.Process(&batch)
 	assert.NoError(t, err)
-	pod, found := result.MetricSets[core.PodKey("ns1", "pod1")]
+	cluster, found := result.MetricSets[core.ClusterKey()]
 	assert.True(t, found)
 
-	m1, found := pod.MetricValues["m1"]
+	m1, found := cluster.MetricValues["m1"]
 	assert.True(t, found)
 	assert.Equal(t, 110, m1.IntValue)
 
-	m2, found := pod.MetricValues["m2"]
-	assert.True(t, found)
-	assert.Equal(t, 222, m2.IntValue)
-
-	m3, found := pod.MetricValues["m3"]
+	m3, found := cluster.MetricValues["m3"]
 	assert.True(t, found)
 	assert.Equal(t, 30, m3.IntValue)
-
-	labelPodName, found := pod.Labels[core.LabelPodName.Key]
-	assert.True(t, found)
-	assert.Equal(t, "pod1", labelPodName)
-
-	labelNsName, found := pod.Labels[core.LabelNamespaceName.Key]
-	assert.True(t, found)
-	assert.Equal(t, "ns1", labelNsName)
 }
