+++ metrics/processors/kubeprocessors/namespace_aggregator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package processors
+package kubeprocessors
 
 import (
 	"github.com/golang/glog"
